+++ poetry/console/commands/search.py
@@ -1,14 +1,16 @@
+from cleo import argument
+from cleo import option
+
 from .command import Command
 
 
 class SearchCommand(Command):
-    """
-    Searches for packages on remote repositories.
 
-    search
-        { tokens* : The tokens to search for. }
-        { --N|only-name : Search only in name. }
-    """
+    name = "search"
+    description = "Searches for packages on remote repositories."
+
+    arguments = [argument("tokens", "The tokens to search for.", multiple=True)]
+    options = [option("only-name", "N", "Search only in name.")]
 
     def handle(self):
         from poetry.repositories.pypi_repository import PyPiRepository