+++ poetry/console/commands/update.py
@@ -1,17 +1,29 @@
+from cleo import argument
+from cleo import option
+
 from .env_command import EnvCommand
 
 
 class UpdateCommand(EnvCommand):
-    """
-    Update dependencies as according to the <comment>pyproject.toml</> file.
 
-    update
-        { packages?* : The packages to update. }
-        { --no-dev : Do not install dev dependencies. }
-        { --dry-run : Outputs the operations but will not execute anything
-                      (implicitly enables --verbose). }
-        { --lock : Do not perform install (only update the lockfile). }
-    """
+    name = "update"
+    description = (
+        "Update dependencies as according to the <comment>pyproject.toml</> file."
+    )
+
+    arguments = [
+        argument("packages", "The packages to update", optional=True, multiple=True)
+    ]
+    options = [
+        option("no-dev", None, "Do not update dev dependencies."),
+        option(
+            "dry-run",
+            None,
+            "Output the operations but do not execute anything "
+            "(implicitly enables --verbose).",
+        ),
+        option("lock", None, "Do not perform operations (only update the lockfile)."),
+    ]
 
     loggers = ["poetry.repositories.pypi_repository"]