+++ poetry/packages/vcs_dependency.py
@@ -62,6 +62,17 @@ class VCSDependency(Dependency):
 
         return "{} {}".format(what, version)
 
+    @property
+    def base_pep_508_name(self):  # type: () -> str
+        requirement = self.pretty_name
+
+        if self.extras:
+            requirement += "[{}]".format(",".join(self.extras))
+
+        requirement += " @ {}+{}@{}".format(self._vcs, self._source, self.reference)
+
+        return requirement
+
     def is_vcs(self):  # type: () -> bool
         return True